Fix ICE in ArgFolder when MIR contains closure types
