+++ tests/ui/const-generics/issues/ice-issue-105238-2.rs
@@ -1,4 +1,4 @@
-//@ known-bug: #105238
+// Regression test for #105238: default type param Ret::R used to ICE in ArgFolder.
 
 #![allow(incomplete_features)]
 #![feature(generic_const_exprs)]
@@ -27,5 +27,7 @@ struct RobinHashTable<
 fn main() {
     use std::mem::size_of;
     println!("{}", size_of::<RobinHashTable<1024>>());
+    //~^ ERROR the size for values of type `CellIdx` cannot be known at compilation time
     println!("{}", size_of::<RobinHashTable<65536>>());
+    //~^ ERROR the size for values of type `CellIdx` cannot be known at compilation time
 }
