+++ tests/ui/const-generics/issues/ice-issue-105238-2.stderr
@@ -1,0 +1,49 @@
+error[E0277]: the size for values of type `CellIdx` cannot be known at compilation time
+  --> $DIR/ice-issue-105238-2.rs:29:30
+   |
+LL |     println!("{}", size_of::<RobinHashTable<1024>>());
+   |                              ^^^^^^^^^^^^^^^^^^^^ doesn't have a size known at compile-time
+   |
+note: required by an implicit `Sized` bound in `RobinHashTable`
+  --> $DIR/ice-issue-105238-2.rs:22:5
+   |
+LL | struct RobinHashTable<
+   |        -------------- required by a bound in this struct
+LL |     const MAX_LENGTH: usize,
+LL |     CellIdx = <Cond<{ MAX_LENGTH < 65535 }, u16, u32> as Ret>::R,
+   |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by the implicit `Sized` requirement on this type parameter in `RobinHashTable`
+help: you could relax the implicit `Sized` bound on `CellIdx` if it were used through indirection like `&CellIdx` or `Box<CellIdx>`
+  --> $DIR/ice-issue-105238-2.rs:22:5
+   |
+LL |     CellIdx = <Cond<{ MAX_LENGTH < 65535 }, u16, u32> as Ret>::R,
+   |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ this could be changed to `CellIdx: ?Sized`...
+LL | > {
+LL |     _idx: CellIdx,
+   |           ------- ...if indirection were used here: `Box<CellIdx>`
+
+error[E0277]: the size for values of type `CellIdx` cannot be known at compilation time
+  --> $DIR/ice-issue-105238-2.rs:31:30
+   |
+LL |     println!("{}", size_of::<RobinHashTable<65536>>());
+   |                              ^^^^^^^^^^^^^^^^^^^^^ doesn't have a size known at compile-time
+   |
+note: required by an implicit `Sized` bound in `RobinHashTable`
+  --> $DIR/ice-issue-105238-2.rs:22:5
+   |
+LL | struct RobinHashTable<
+   |        -------------- required by a bound in this struct
+LL |     const MAX_LENGTH: usize,
+LL |     CellIdx = <Cond<{ MAX_LENGTH < 65535 }, u16, u32> as Ret>::R,
+   |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by the implicit `Sized` requirement on this type parameter in `RobinHashTable`
+help: you could relax the implicit `Sized` bound on `CellIdx` if it were used through indirection like `&CellIdx` or `Box<CellIdx>`
+  --> $DIR/ice-issue-105238-2.rs:22:5
+   |
+LL |     CellIdx = <Cond<{ MAX_LENGTH < 65535 }, u16, u32> as Ret>::R,
+   |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ this could be changed to `CellIdx: ?Sized`...
+LL | > {
+LL |     _idx: CellIdx,
+   |           ------- ...if indirection were used here: `Box<CellIdx>`
+
+error: aborting due to 2 previous errors
+
+For more information about this error, try `rustc --explain E0277`.
