+++ tests/ui/const-generics/issues/ice-issue-93182.rs
@@ -1,16 +1,14 @@
-//@ known-bug: #93182
+// Regression test for #93182: trait const default CONST used to ICE in ArgFolder.
+
+#![allow(incomplete_features)]
 #![feature(generic_const_exprs)]
 
-// this causes an ICE!!!
 pub const CONST: usize = 64;
 pub trait Tr<const S: usize = CONST>: Foo<A<S>> {}
 
-// no ICE
-// pub trait Digest<const S: usize = 64>: FromH<[u8; S]> {}
+struct St();
 
-struct St ();
-
-struct A<const S: usize> ([u8; S]);
+struct A<const S: usize>([u8; S]);
 
 pub trait Foo<T> {
     fn foo(_: T);
@@ -24,6 +22,7 @@ impl<const S: usize> Foo<A<S>> for St {
 
 pub trait FooBar {
     type Tr: Tr;
+    //~^ ERROR mismatched types
 }
 
 pub fn main() {}
